adding cards from for each pizza builder section
