+++ src/PizzaBuilder/components/AccordionContents/AccordionContent.js
@@ -1,25 +1,54 @@
 import React from 'react';
+import axios from 'axios';
 import ContentCard from './ContentCard';
+import { connect } from 'react-redux';
 
 class AccordionContent extends React.Component{
-    // state=
-    // {
-    //     cards: []
-    // }
+    constructor(props){
+        super(props);
+    }
+
+    state=
+    {
+        cards: []
+    }
     
     componentDidMount(){
-        
-        // const { cards } =  this.props.cards;
-        // this.setState({cards});
-        // console.log(cards);
-           
+        // axios.get('http://pizzadeploy-env.dn37p3zqw3.ap-southeast-2.elasticbeanstalk.com/ingredient/SAUCES')
+        //     .then(response => {
+        //         const contents = response.data.data;
+        //         this.setState({contents});
+        //    })
+
+        switch(this.props.content){
+            case 'TOPPINGS':
+                let TOPPING, MEAT , VEG = [];
+                MEAT = this.props.MEATS;
+                VEG = this.props.VEGGIGS;
+                TOPPING =  [...MEAT, ...VEG]
+                this.setState({cards : TOPPING})
+                break;
+            case 'SAUCES':
+                this.setState({cards : this.props.SAUCES})
+                break;
+            case 'CHEESES':
+                this.setState({cards: this.props.CHEESES})
+                break;
+            default:
+                this.setState({cards: [] })
+            
+        }
+
     } 
 
     render(){
+        
         return(
             <section className='content__flex'>
             {
+                
                 this.state.cards.map( card => (
+
                 <ContentCard
                 key={card.id}
                 name={card.IngredientName}
@@ -38,4 +67,15 @@ class AccordionContent extends React.Component{
     }
 }
 
-export default AccordionContent;
+const mapStateToProps = (state) => ({
+    MEATS: state.ingredient.MEATS,
+    CHEESES: state.ingredient.CHEESES,
+    SAUCES: state.ingredient.SAUCES,
+    VEGGIGS: state.ingredient.VEGGIGS,
+    isLoading: state.ingredient.isLoading,
+    errorMessage:state.ingredient.errorMessage,
+  });
+
+  export default connect(mapStateToProps)(AccordionContent);
+
+
